# api/swagger.yaml
definitions:
  model.SlowestQuery:
    properties:
      created_at:
        type: string
      id:
        type: integer
      query:
        type: string
      statement:
        type: string
      time_spent:
        type: integer
      updated_at:
        type: string
    required:
    - query
    - statement
    - time_spent
    type: object
info:
  contact: {}
paths:
  /api/v1/query:
    delete:
      parameters:
      - description: id of the query
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
      summary: Delete single query entity
      tags:
      - query
    get:
      parameters:
      - description: id of the ad
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SlowestQuery'
        "500":
          description: Internal Server Error
      summary: Get single query entity
      tags:
      - query
    post:
      consumes:
      - application/json
      responses:
        "201":
          description: Created
        "500":
          description: Internal Server Error
      summary: Creates a single Query entity
      tags:
      - query
    put:
      parameters:
      - description: id of the query
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
      summary: Update single query entity
      tags:
      - query
  /faker:
    get:
      parameters:
      - description: number of queries to generate
        in: query
        name: num
        type: integer
      responses:
        "200":
          description: OK
      summary: Fake Query entities
      tags:
      - faker
swagger: "2.0"